--- lib/poto/download.rb
@@ -1,21 +1,0 @@
-require "open-uri"
-require "poto/ext/open_uri"
-
-module Poto
-  class Download
-    attr_reader :file, :uri
-
-    def initialize(file:, uri:)
-      @file, @uri = file, uri
-    end
-
-    def call
-      open(uri) do |src|
-        file.write(src.read)
-      end
-
-      file.close
-      file.path
-    end
-  end
-end
